% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustic_features.R
\name{acoustic_features}
\alias{acoustic_features}
\title{Acoustic features of wave file}
\usage{
acoustic_features(file, base_output_directory)
}
\arguments{
\item{fwave}{}
}
\value{

}
\description{
Acoustic features of wave file
}
