% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bat_detection.R
\name{bat_detection}
\alias{bat_detection}
\title{Detect bat calls from spectrogram peaks}
\usage{
bat_detection(fwave)
}
\arguments{
\item{fwave}{wave file}
}
\value{

}
\description{
Detect bat calls from spectrogram peaks
}
