% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_indices.R
\name{calculate_indices}
\alias{calculate_indices}
\title{Calculate analysis programs from within R}
\usage{
calculate_indices(file, base_output_directory)
}
\arguments{
\item{file}{wave file name}

\item{base_output_directory}{place to put AP output}
}
\value{

}
\description{
Calculate analysis programs from within R
}
