% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_indices.R
\name{clean_indices}
\alias{clean_indices}
\title{Clean AP output data sets}
\usage{
clean_indices(output_directory, file_name, ind)
}
\arguments{
\item{ind}{Three digit code of acoustic indices}
}
\value{

}
\description{
Clean AP output data sets
}
