% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_plot.R
\name{tile_plot}
\alias{tile_plot}
\title{Easy tile plot}
\usage{
tile_plot(
  x,
  chosen_indices = c("ACI", "ENT", "BGN", "CVR", "DIF", "EVN", "PMN", "R3D", "RHZ",
    "RNG", "RPS", "RVT", "SPT", "SUM"),
  sample_length,
  mytitle = F
)
}
\arguments{
\item{x}{data frame}

\item{chosen_indices}{can be single or multiple indices as the three digit code}

\item{sample_length}{max time length (seconds)}

\item{mytitle}{optional}
}
\value{

}
\description{
Easy tile plot
}
